Yuchen's feedback 2020/09/08
The No.1 thing to figure out is the the consistency between the terms CDN vs
gateway cache. We also need to define them so that the readers of this
RFC know if the services they are running are relevant.

The biggest change in my revision is that whether to preseve
CDN-Control header after a gateway proxy uses it.

Others changes are more or less fixing typos and wording.